move generic code to dedicated files with build tag
